More effort in providing gapless joining even if accessd out of sequence.
Some music player combinations look ahead to the next file which results
in non-gapless joining from the previous file, because we already planted
a version of that file in the cache.

Once we reach the end of the previous file we now create a fresh version
of that file to properly gapless join regardless.

This is still best effort though: if the next file is still opened in
the filesystem, we don't attempt to have two concurrent versions of the
same file open (TODO).