Small fixes and make compatible with current microhttpd
Turns out the cut-off version of the new feature of microhttpd is
slightly different, so adapt that.

Fix an oversight of a fully-qualified write() loop - it didn't
actually do what it meant to do.

Makefile: make CXX replaceable by environment variable. Create
directory first before installing in case it is not there yet.